Implement all the simple row-column fishes.
This implements all the following fishes, as long as the base houses
are rows and the covering houses are columns, or vice versa (in the
terminology of http://hodoku.sourceforge.net/en/tech_fishg.php):
  - x-wings (2-fish)
  - swordfish (3-fish)
  - jellyfish (4-fish)

Details:

All these simple row-column fishes are trivial to implement,
by calling the already-existing function eliminate_candidates()
on the board transformed by transposing the columns with the digits,
using only the rows and columns (not blocks) as the groups.

(There is no need to do the same with *rows*-transposed-with-digits,
since every k-fish that would be found in that way
is found as a complementary (9-k)-fish on cols-transposed-with-digits.)

This allows many more puzzles to be solved with less backtracking
or no backtracking than before.  For example, the default (Everest)
puzzle now takes only 36 calls to solve() instead of 59,
and now only one of the 50 example boards from projecteuler
still requires backtracking (4 of them required it before).
In bash:
  for line in $(cat collections/project_euler.txt ); do echo $line; python2 sudoku.py $line; done | grep num_solves | cat -n | grep -v "num_solves = 1"
Output is:
  42  num_solves = 2
Note: the reason #42 can't be done without backtracking is that it
requires pointing or claiming, which is currently not implemented
(contrary to what README.md seems to imply).

Note: The following comment which is still in README.md is evidently incorrect:
  "All these strategies [...] are, in essence, _search and backtracking_
   strategies in disguise, no more clever than brute forcing through all
   possible combinations."
As is shown in this commit, the basic fishes are actually
yet another case of the pigeonhole principle (aka "the Rule"),
and, as such, require no search-and-backtracking.
In fact, I would say that even the more exotic fishes described on the
hodoku page are still cases of the pigeonhole principle,
but they are not as easily expressed in terms of the already-written function
eliminate_candidates(), so they are not implemented here.